# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/{username}": {
      "get": {
        "operationId": "UsersController_getUser",
        "summary": "Information about a user via username",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
    "/requests": {
      "post": {
        "operationId": "RequestsController_createRequest",
        "summary": "Create a request to management",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRequest"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ]
      },
      "get": {
        "operationId": "RequestsController_getAll",
        "summary": "Get all requests",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffRequest"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ]
      }
    },
    "/requests/{id}": {
      "get": {
        "operationId": "RequestsController_getOne",
        "summary": "Get one requests",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRequest"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ]
      },
      "delete": {
        "operationId": "RequestsController_delete",
        "summary": "Delete a request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRequest"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ]
      }
    },
    "/requests/attachment/{attachmentId}": {
      "get": {
        "operationId": "RequestsController_getAttachment",
        "summary": "Get attachment information",
        "parameters": [
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "name": "name",
                  "filename": "filename",
                  "text": "content"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ]
      }
    },
    "/admin/attendance/{username}": {
      "get": {
        "operationId": "AdminController_findUserAttendance",
        "summary": "Attendance of a user",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "description": "Select date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "description": "Select date to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Select a specific date",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Attendance"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/attachment/{attachmentId}": {
      "get": {
        "operationId": "AdminController_findAttendanceAttachment",
        "summary": "Get an attachment.",
        "parameters": [
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "name": "name",
                  "filename": "filename",
                  "text": "content"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/requests": {
      "get": {
        "operationId": "AdminController_getAllReuests",
        "summary": "All requests from all users",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffRequest"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/requests/{id}": {
      "get": {
        "operationId": "AdminController_getRequest",
        "summary": "Get a specific request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRequest"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/requests/{id}/status": {
      "patch": {
        "operationId": "AdminController_changeStatus",
        "summary": "Change request status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequestStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRequest"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/schedule/{userId}": {
      "post": {
        "operationId": "AdminController_createStaffSchedule",
        "summary": "Create schedules for user",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateScheduleDto"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateScheduleDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/ip-addr": {
      "get": {
        "operationId": "AdminController_getIP",
        "summary": "Get all IP addresses",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRequest"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminController_registerIP",
        "summary": "Register IP Address",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIPDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateIPDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/ip-addr/{id}": {
      "put": {
        "operationId": "AdminController_editIP",
        "summary": "Update name or enable/disable an IP Address",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditIPDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditIPDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteIP",
        "summary": "Delete IP Address",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/user/{id}": {
      "delete": {
        "operationId": "AdminController_deleteUser",
        "summary": "Delete User Account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/feedback-analytics": {
      "get": {
        "operationId": "AdminController_fetchAllFeedback",
        "summary": "Fetch feedback with parameters",
        "parameters": [
          {
            "name": "user",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "industry",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "company",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "feedbacks",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "totalFeedbacks": 3,
                  "industryAnalytics": [
                    {
                      "industry": {
                        "id": 1,
                        "name": "Something",
                        "description": "some thing some thing"
                      },
                      "status": {
                        "ACCEPTED": 1,
                        "ALMOST_ACCEPTED": 0,
                        "UNDECIDED": 1,
                        "ALMOST_REJECTED": 0,
                        "REJECTED": 1
                      },
                      "feedbacks": []
                    }
                  ],
                  "total": {
                    "ACCEPTED": 1,
                    "ALMOST_ACCEPTED": 0,
                    "UNDECIDED": 1,
                    "ALMOST_REJECTED": 0,
                    "REJECTED": 1
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Attendance API",
    "description": "API documentation for attendance backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "MANAGEMENT",
              "MARKETING",
              "DEVELOPMENT",
              "SALES",
              "RND"
            ]
          }
        },
        "required": [
          "name",
          "username",
          "email",
          "password",
          "role"
        ]
      },
...
      "Attendance": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "IN",
              "OUT",
              "PAUSE",
              "RESUME"
            ]
          },
          "filename": {
            "type": "string"
          },
          "date": {
            "type": "number"
          },
          "id": {
            "type": "number"
          }
        },
        "required": [
          "userId",
          "status",
          "filename",
          "date",
          "id"
        ]
      },
      "CreateRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "content",
          "filename"
        ]
      },
      "StaffRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "userId": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "PENDING",
              "REJECTED",
              "CONSULTATION_REQUIRED"
            ]
          }
        },
        "required": [
          "title",
          "content",
          "filename",
          "userId",
          "status"
        ]
      },
      "UpdateRequestStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "PENDING",
              "REJECTED",
              "CONSULTATION_REQUIRED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          },
          "day": {
            "type": "string"
          },
          "uniform": {
            "type": "boolean"
          }
        },
        "required": [
          "startTime",
          "endTime",
          "day",
          "uniform"
        ]
      },
      "CreateIPDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "address",
          "name"
        ]
      },
      "EditIPDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "address",
          "name",
          "enabled"
        ]
      },
...
    }
  }
}